requirements.txt
Flask==2.0.3
SQLAlchemy==1.4.31
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Caching==2.0.2
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
fakeredis==2.20.0
factory-boy==3.3.0
redis==5.0.1
psycopg2-binary==2.9.3
werkzeug==2.0.3
gunicorn==20.1.0